shared/requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
